Refactored a bunch of MainScreenController into GameLoop < layout, organisation
