+++ src/main/java/nl/tudelft/fishy/factories/LosingScreenEventHandlerFactory.java
@@ -2,6 +2,7 @@ package nl.tudelft.fishy.factories;
 
 import nl.tudelft.fishy.interfaces.LosingScreenEventHandlerFactoryInterface;
 import nl.tudelft.fishy.Game;
+import nl.tudelft.fishy.GameLoop;
 import nl.tudelft.fishy.controllers.MainScreenController;
 import javafx.event.EventHandler;
 import javafx.scene.input.MouseEvent;
@@ -70,7 +71,7 @@ public final class LosingScreenEventHandlerFactory implements
 
     		Game.resetPlayerFishSize();
     		MainScreenController.playerFish.setHasLance(false);
-    		MainScreenController.setBossMode(false);
+    		GameLoop.setBossMode(false);
 
     		MainScreenController.playerFish.setHasLance(false);
